fixed an annoying bug
When running a bot as a module the plugins were being registered by the
framework each time they were changed as a new plugin. So if you were to
have a plugin hooked by .foo that returned 'bar' and then changed the
return value to 'baz' the plugin would return 'bar' and then 'bad', i.e.
it was loaded twice. Now that's fixed.