fixing a bug that caused duplicate plugins upon reload
This is the same annoying bug that popped up before. The solution is to
clear the loaded plugins list and then repopulate it with the latest
list. Now this is made more robust by only generating a list of
"filtered" plugins when the functions that are hooked by plugins are
actually found in the file we're reloading.