fixing a single-module reload ug, updating examples
This update includes a hack that the fixes the filtering out of single-module
bot source files. Due to the fact that the modules reloaded from sys.path
include some system-specific modules that should not be reloaded we limit the
modules to modules containing the abs_path of the bot instance as defined at
runtime.

However in the case of a bot that is a single module, this would mean that we
would ignore that candidate file for reload as it would not necessarily contain
the abs_path. In order to circumvent this issue we rely on sys._current_stack
and inspect to pull the frame that the bot instance was instantiated in and
thereby extract the filename it resides in.

Also this commit updates the README and includes another simple example bot.